make distence function, end training

+++ src/orfium_audio/data.py
@@ -1,5 +1,6 @@
 from pathlib import Path
 
+import numpy as np
 import polars as pl
 import torch
 from torch.utils.data import DataLoader, TensorDataset
@@ -13,7 +14,7 @@ class Dataset:
             tabular_dataset = tabular_dataset.drop(drop_columns)
 
         self.tensor_data = TensorDataset(
-            torch.from_numpy(tabular_dataset.collect().to_numpy())
+            torch.from_numpy(tabular_dataset.collect().to_numpy().astype(np.float32))
         )
 
     def get_loader(self, batch_size: int = 512) -> DataLoader:
